move emailEmbedSchema to own file
