+++ libs/newsletters-data-client/src/lib/emailEmbedSchema.ts
@@ -1,0 +1,11 @@
+import { z } from 'zod';
+
+export const emailEmbedSchema = z.object({
+	name: z.string(),
+	title: z.string(),
+	description: z.string().optional(),
+	successHeadline: z.string(),
+	successDescription: z.string(),
+	hexCode: z.string(),
+	imageUrl: z.string().optional(),
+});
