+++ libs/newsletters-data-client/src/lib/schema-helpers.ts
@@ -3,12 +3,3 @@ import { z } from 'zod';
 export const nonEmptyString = () =>
 	z.string().min(1, { message: 'Must not be empty' });
 
-export const emailEmbedSchema = z.object({
-	name: z.string(),
-	title: z.string(),
-	description: z.string().optional(),
-	successHeadline: z.string(),
-	successDescription: z.string(),
-	hexCode: z.string(),
-	imageUrl: z.string().optional(),
-});
